feat: Complete Fresh Greens e-Market platform
- Spring Boot 4.0.3 with Firebase Auth, MySQL/Hibernate, Razorpay
- HTML/Bootstrap 5/jQuery frontend (10 pages)
- Session persistence via Spring Security context
- In-memory caching for development
- Remove secrets from tracking (firebase credentials, copilot config)

+++ app/src/main/java/com/freshgreens/app/service/OrderService.java
@@ -74,7 +74,7 @@ public class OrderService {
 
         // Create Razorpay order
         JSONObject rzpOrderRequest = new JSONObject();
-        rzpOrderRequest.put("amount", grandTotal.multiply(BigDecimal.valueOf(100)).intValue()); // Razorpay expects paise
+        rzpOrderRequest.put("amount", grandTotal.multiply(BigDecimal.valueOf(100)).longValue()); // Razorpay expects paise
         rzpOrderRequest.put("currency", "INR");
         rzpOrderRequest.put("receipt", orderNumber);
